Added multiple server capabilities

+++ lib/cogs/minecraftController.py
@@ -14,6 +14,8 @@ bot_dir=r"C:\Users\Admin\Desktop\Trea"
 class MinecraftController(Cog):
     def __init__(self, bot):
         self.bot = bot
+        self.activeServer = None
+        self.activeServerName = None
 
     @Cog.listener()
     async def on_ready(self):
@@ -25,20 +27,28 @@ class MinecraftController(Cog):
 
     @command(aliases=['launch', 'launchserver'])
     async def launchServer(self, ctx, server):
-        await ctx.send(f'Attempting to launch server: {server}')
-        server_dir = os.path.join('MinecraftServers', f'{server}')
-        assert os.path.isdir(server_dir)
-        env = dict(os.environ)
-        env['JAVA_OPTS'] = 'foo'
-        self.activeServer = subprocess.Popen(['java', '-Xmx4096M', '-Xms4096M', '-jar', f'{server}-launch.jar'], env=env, cwd=server_dir)
+        if (self.activeServer != None):
+            await ctx.send(f'A server is already running')
+        else:
+            await ctx.send(f'Attempting to launch server: {server}')
+            server_dir = os.path.join('MinecraftServers', f'{server}')
+            assert os.path.isdir(server_dir)
+            env = dict(os.environ)
+            env['JAVA_OPTS'] = 'foo'
+            self.activeServer = subprocess.Popen(['java', '-server', '-Xmx6G', '-Xms6G', '-jar', f'{server}-launch.jar'], env=env, cwd=server_dir)
+            self.activeServerName = server
 
-
-    @command()
+    @command(aliases=['kill'])
     async def killServer(self, ctx):
-        await ctx.send(f'Attempting to kill server.')
-        self.activeServer.terminate()
+        if (self.activeServer != None):
+            await ctx.send(f'Attempting to kill server: {self.activeServerName}')
+            self.activeServer.terminate()
+            self.activeServer = None
+            self.activeServerName = None
+        else:
+            await ctx.send(f'There is no server running')
 
-    @command()
+    @command(aliases=['list', 'listservers'])
     async def listServers(self, ctx):
         dirs = os.listdir('MinecraftServers')
         
@@ -52,6 +62,14 @@ class MinecraftController(Cog):
                 await ctx.send(f'{d} directory accessed.')
             except IOError:
                 print("File not accessible")
+
+    @command(aliases=['active', 'activeservers'])
+    async def activeServers(self, ctx):
+        if (self.activeServer != None):
+            await ctx.send(f'The current active server is: {self.activeServerName}')
+        else:
+            await ctx.send(f'There is no server running')
+        
 
 def setup(bot):
     bot.add_cog(MinecraftController(bot))
